refactor(workspace): rename files
